Fix all clippy warnings across the workspace
- gruel-parser: remove redundant .clone() on Copy type (Ident)
- gruel-air/inference: allow too_many_arguments on ConstraintGenerator::new
- gruel-air/sema: collapse nested if-let chains into && patterns (4 sites)
- gruel-air/sema: allow too_many_arguments on analyze_enum_struct_variant
- gruel-codegen-llvm: remove redundant .into() on AggregateValueEnum (2 sites)
- gruel-codegen-llvm: remove unnecessary u64 cast on array length

+++ crates/gruel-codegen-llvm/src/types.rs
@@ -84,7 +84,7 @@ pub fn type_byte_size(ty: Type, type_pool: &TypeInternPool) -> u64 {
         }
         TypeKind::Array(id) => {
             let (elem_ty, len) = type_pool.array_def(id);
-            type_byte_size(elem_ty, type_pool) * len as u64
+            type_byte_size(elem_ty, type_pool) * len
         }
         TypeKind::Enum(id) => {
             let def = type_pool.enum_def(id);
